clustering/requirements.txt
mysql-connector-python==8.0.33
numpy==1.24.3
scikit-learn==1.3.0

